First Update: Taking input from text file.
In this code, 2 main classes are created that is TV Channel and Family. And for each member of the family data is taken as input from the text file. Similarly for the shows in TV Channel, input is taken from the text file. Created a list individually for every shows which contains the names of family members whose favorite show it is.